requirements.txt
argparse==1.2.1
beautifulsoup4==4.3.2
decorator==3.4.0
mock==1.0.1
networkx==1.9.1
nltk==3.0.1
nose==1.3.4
numpy==1.9.2
pyparsing==2.0.3
python-dateutil==2.4.1
python-magic==0.4.6
pytz==2014.10
requests==2.5.3
six==1.9.0
wsgiref==0.1.2
